+++ web/documentserver-example/nodejs/helpers/wopi/utils.js
@@ -165,8 +165,8 @@ const getEditNewText = function getEditNewText(ext) {
       return 'Presentation';
     case 'pdf':
       return 'PDF form';
-    case 'visio':
-      return 'Visio';
+    case 'diagram':
+      return 'Diagram';
     default:
       return ext.toUpperCase();
   }